backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "fastapi[standard]>=0.115.12",
    "isort>=6.0.1",
    "langchain>=0.3.24",
    "langchain-chroma>=0.2.3",
    "langchain-community>=0.3.23",
    "langchain-huggingface>=0.1.2",
    "marker-pdf>=1.6.2",
    "pandas>=2.2.3",
    "torch>=2.7.0",
    "transformers>=4.51.3",
]
